auto to endentity

+++ src/main/resources/db/migration/V10__Create_task_status_Table.sql
@@ -1,5 +1,5 @@
 create table task_status (
-    id int8 not null,
+    id bigserial not null,
     name varchar(255),
     status varchar(255),
     primary key (id)
